requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
cryptography
Pillow
